Fix A* heuristic calculation
Turns out that the problem was caused by calcuating the distance from
the current node to the explored node, rather than the distance from the
end node to the explored node... this essentially nuked the A*
heuristic. Because of the way Java priority queues are implemented, this
caused the algorithm to revert to a scanline search method instead of
going to something akin to Dijkstra's.