Fix pathfinding efficiency
That was weird... now it checks to see if nodes are in the open list
before they get pushed in. If one with a less than or equal f value is found, the
insert operation is aborted. If one with a greater f value is found,
it's replaced. If one isn't found at all, the node is just flat out
inserted.