Fix issue #12
Turns out that the pathfinding completion check was parenting the end
node to curNode's parent when it should've been parenting to curNode
itself. This has been fixed, a test case has been added to verify it,
and the backtracking code has been simplified (it was more complex than
it needed to be, although it wasn't the problem).