Implement hash table caching and linkage preseravtion + tests
This method of handling the database is complex under the hood, but
extremely easy to work with. It relies on maintaing a hash table of
nodes associated with their IDs. When a request for a node comes in, the
database first checks the cache to see if the node has been already
loaded. If it has, the cache is queried instead and the node is
returned. If the node has NOT been loaded, the database is queried
instead, and all relevant linkages are made (in both directions). Any
time an insert operation is made, the cache and database are updated
simultaneously, and the same goes for deletions. When the database first
starts up, it immediately calls reloadCache(), which loads all nodes
into the cache and establishes all links between them. This way, the
database user has minimal involvement in how the database actually
operates - the goal of this is to make the database as easy to use and
resilient as possible.