Fix testProviderOperations
The unit test is fixed, but some serious overhauls on providers are
needed. Right now the database generates them on the fly, and that just
isn't going to work in the long run, for the same reason that nodes have
to be stored in a separate cache. The next few commits will address this
issue.